fibon

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,9 +11,9 @@
 int main(void)
 {
 	int i = 0;
-	int a = 1;
-	int b = 2;
-	int sum = 0;
+	long int a = 1;
+	long int b = 2;
+	long int sum = 0;
 
 	printf("%i, %i, ", a, b);
 	for (i = 1; i < 50; i++)
